feat: Add new cup component for wrapping designs
The new cup component, "NewCup", is added to the Experience component to allow users to wrap designs around the cup. The previous cup component, "WrappableCup", is commented out and replaced with the new component. This change improves the functionality of the application by providing a more flexible and customizable cup wrapping feature.